+++ Tweaks/UiAdjustment/OpenGlamourDresserToCurrentJob.cs
@@ -1,4 +1,4 @@
-﻿using FFXIVClientStructs.FFXIV.Component.GUI;
+using FFXIVClientStructs.FFXIV.Component.GUI;
 using SimpleTweaksPlugin.Events;
 using SimpleTweaksPlugin.TweakSystem;
 
@@ -7,7 +7,7 @@ namespace SimpleTweaksPlugin.Tweaks.UiAdjustment;
 [TweakName("Open Glamour Dresser to Current Job")]
 [TweakAuthor("MidoriKami")]
 [TweakDescription("Sets the job selection dropdown to your current job when opening the Glamour Dresser.")]
-[TweakReleaseVersion(UnreleasedVersion)]
+[TweakReleaseVersion("1.9.0.0")]
 public unsafe class OpenGlamourDresserToCurrentJob : UiAdjustments.SubTweak {
     [AddonPreSetup("MiragePrismPrismBox")]
     private void OnMiragePrismBoxOpen(AtkUnitBase* atkUnitBase) {